+++ generators/app/project-types.js
@@ -7,10 +7,10 @@ module.exports = [
     name: '02 - Micro-Service',
     value: '02-microservice'
   },
-  {
-    name: '03 - Hybrid App',
-    value: '03-hybrid-app'
-  },
+  // {
+  //   name: '03 - Hybrid App',
+  //   value: '03-hybrid-app'
+  // },
   // {
   //   name: '04 - Clean Architect Project',
   //   value: '04-clean-architect'