+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Change Log
 ## 1.1.0
-- Add project type `microservice` & `hybrid-app`
-- Automatic add async command when create module
+- Add project type `microservice`
 - Update use openapi/asyncapi for API Document (solution for both of standalone app & micoservice)
 - Bugs fixing
 ## 1.0.1